An S4 class to represent technology

An S4 class to represent technology

Create object of class `technology`.

Usage

newTechnology(
  name = "",
  desc = "",
  input = data.frame(),
  output = data.frame(),
  group = data.frame(),
  aux = data.frame(),
  units = data.frame(),
  cap2act = as.numeric(1),
  geff = data.frame(),
  ceff = data.frame(),
  aeff = data.frame(),
  af = data.frame(),
  afs = data.frame(),
  weather = data.frame(),
  capacity = data.frame(),
  invcost = data.frame(),
  fixom = data.frame(),
  varom = data.frame(),
  olife = data.frame(),
  region = character(),
  start = data.frame(),
  end = data.frame(),
  timeframe = character(),
  fullYear = TRUE,
  optimizeRetirement = FALSE,
  misc = list()
)

# S4 method for class 'technology'
update(object, ...)

Arguments

object

object of class technology

...

slot-names with data to update (see `newTechnology`)

Value

an object of class technology.

Slots

name

character, name of technology used in sets

desc

character, optional description of the technology for reference.

input

main (non-auxilary) input commodities, data.frame with columns:

comm

character, name of input commodity

unit

character, unit of the input commodity

group

character, name of input group for input commodities-substitutes

combustion

numeric, combustion factor from 0 to 1 (default 1) to calculate emissions from fuels combustion (commodities intermediate consumption, more broadly)

output

main (non-auxilary) output commodities, data.frame with columns:

comm

character, name of output commodity

unit

character, unit of the output commodity

group

character, name of output group for output commodities-substitutes

aux

auxilary commodities, both input and output, data.frame with columns:

acomm

character, name of auxilary commodity

unit

character, unit of the auxilary commodity

units

key units of the process activity and capacity. data.frame with columns:

capacity

character, name of capacity unit

use

character, name of use unit

activity

character, name of activity unit

costs

character, name of cost unit

group

optional details for groups, data.frame with columns:

group

character, name of group

desc

character, optional description of the group

unit

character, optional unit of the group

cap2act

numeric. Capacity to activity ratio. Default 1. Specifies how much product (output) will be produced per unit of capacity.

geff

input-group efficiency parameter, data.frame with columns:

region

character, name of region for which the parameter will be applied, NA for every region

year

integer, year for which the parameter will be applied, NA for every year

slice

character, name of slice for which the parameter will be applied, NA for every slice

group

character, name of group for which the parameter will be applied, required

ginp2use

numeric, group input to use ratio, NA means no changes of default value

ceff

commodity-related efficiency parameters, data.frame with columns:

region

character, name of region for which the parameter will be applied, NA for every region

year

integer, year for which the parameter will be applied, NA for every year

slice

character, name of slice for which the parameter will be applied, NA for every slice

comm

character, name of commodity for which the parameter will be applied, required

cinp2use

numeric, commodity input to use parameter, NA means default value

use2cact

numeric, use to commodity activity parameter, NA means default value

cact2cout

numeric, commodity activity to commodity output parameter, NA means default value (1)

cinp2ginp

numeric, commodity input to group input parameter, NA means default value (1)

share.lo

numeric, lower bound on a share of commodity within a group, NA means default value(0) or fixed value if specified

share.up

numeric, upper bound on a share of commodity within a group, NA means default value(1) or fixed value if specified

share.fx

numeric, fixed share of commodity within a group, NA ignored

afc.lo

numeric, lower bound on the physical value of the commodity, NA ignored

afc.up

numeric, upper bound on the physical value of the commodity, NA ignored

afc.fx

numeric, fixed physical value of the commodity, NA ignored

aeff

auxilary-commodity-related efficiency parameters, data.frame with columns:

region

character, name of region for which the parameter will be applied, NA for every region

year

integer, year for which the parameter will be applied, NA for every year

slice

character, name of slice for which the parameter will be applied, NA for every slice

acomm

character, name of auxilary commodity for which the parameter will be applied, required

cinp2ainp

numeric, main commodity input to auxilary commodity input parameter, NA ignored

cinp2aout

numeric, main commodity input to auxilary commodity output parameter, NA ignored

cout2ainp

numeric, main commodity output to auxilary commodity input parameter, NA ignored

cout2aout

numeric, main commodity output to auxilary commodity output parameter, NA ignored

act2ainp

numeric, technology activity to auxilary commodity input parameter, NA ignored

act2aout

numeric, technology activity to auxilary commodity output parameter, NA ignored

cap2ainp

numeric, technology capacity to auxilary commodity input parameter, NA ignored

cap2aout

numeric, technology capacity to auxilary commodity output parameter, NA ignored

ncap2ainp

numeric, technology new capacity to auxilary commodity input parameter, NA ignored

ncap2aout

numeric, technology new capacity to auxilary commodity output parameter, NA ignored

stg2ainp

(!!!not implemented!!!) numeric, technology storage level to auxilary commodity input parameter, NA ignored

stg2aout

(!!!not implemented!!!) numeric, technology storage level to auxilary commodity output parameter, NA ignored

sinp2ainp

(!!!not implemented!!!) numeric, technology storage input to auxilary commodity input parameter, NA ignored

sinp2aout

(!!!not implemented!!!) numeric, technology storage input to auxilary commodity output parameter, NA ignored

sout2ainp

(!!!not implemented!!!) numeric, technology storage output to auxilary commodity input parameter, NA ignored

sout2aout

(!!!not implemented!!!) numeric, technology storage output to auxilary commodity output parameter, NA ignored

af

timeslice-level availability factor parameters, data.frame with columns:

region

character, name of region for which the parameter will be applied, NA for every region

year

integer, year for which the parameter will be applied, NA for every year

slice

character, name of slice for which the parameter will be applied, NA for every time-slice of the technology timeframe

af.lo

numeric, lower bound on the availability factor, NA means default value (0) or fixed value if specified

af.up

numeric, upper bound on the availability factor, NA means default value (1) or fixed value if specified

af.fx

numeric, fixed availability factor, NA ignored

rampup

numeric, ramp-up time constraint, NA ignored

rampdown

numeric, ramp-down time constraint, NA ignored

afs

timeframe level (or group of slices) availability factor constraints, data.frame with columns:

region

character, name of region for which the parameter will be applied, NA for every region

year

integer, year for which the parameter will be applied, NA for every year

slice

character, name of the timeframe for which the parameter will be applied, requires, NA will return an error

afs.lo

numeric, lower bound on the availability factor for the timeframe, NA means default value (0) or fixed value if specified

afs.up

numeric, upper bound on the availability factor for the timeframe, NA means default value (1) or fixed value if specified

afs.fx

numeric, fixed availability factor for the timeframe, NA ignored

weather

external ("weather") factors affecting the technology availability paramaters, data.frame with columns:

weather

character, name of the applied weather factor, required

comm

character, name of the commodity with specified `afc.*` affected by the weather factor, required

wafc.lo

numeric, multiplying coefficient to the lower bound on the commodity availability parameter `afc.lo`, NAs ignored of fixed values used

wafc.up

numeric, multiplying coefficient to the upper bound on the commodity availability parameter `afc.up`, NAs ignored of fixed values used

wafc.fx

numeric, multiplying coefficient to the fixed value of the commodity availability parameter `afc.fx`, NAs ignored of fixed values used

waf.lo

numeric, multiplying coefficient to the lower bound on the availability factor parameter `af.lo`, NAs ignored of fixed values used

waf.up

numeric, multiplying coefficient to the upper bound on the availability factor parameter `af.up`, NAs ignored of fixed values used

waf.fx

numeric, multiplying coefficient to the fixed value on the availability factor parameter `af.fx`, NAs ignored of fixed values used

wafs.up

numeric, multiplying coefficient to the upper bound on the availability factor parameter `afs.up`, NAs ignored of fixed values used

wafs.lo

numeric, multiplying coefficient to the lower bound on the availability factor parameter `afs.lo`, NAs ignored of fixed values used

wafs.fx

numeric, multiplying coefficient to the fixed value on the availability factor parameter `afs.fx`, NAs ignored of fixed values used

fixom

fixed operational and maintenance cost (per unit of capacity a year), data.frame with columns:

region

character, name of the region for which the parameter will be applied, NA for every region

year

integer, year for which the parameter will be applied, NA for every year

fixom

numeric, fixed operational and maintannance cost, NA means default value (0)

varom

variable operational and maintenance cost (per unit of activity or commodity), data.frame with columns:

region

character, name of the region for which the parameter will be applied, NA for every region

year

integer, year for which the parameter will be applied, NA for every year

slice

character, name of the time-slice or (grand-)parent timeframe for which the parameter will be applied, NA for every time-slice of the technology timeframe

varom

numeric, variable operational and maintannance cost per unit of activity, NA means default value (0)

comm

character, name of the commodity for which the parameter will be applied, required for `cvarom` parameter

cvarom

numeric, variable operational and maintannance cost per unit of commodity, NA means default value (0)

acomm

character, name of the auxilary commodity for which the `avarom` will be applied, required for `avarom` parameter

avarom

numeric, variable operational and maintannance cost per unit of auxilary commodity, NA means default value (0)

invcost

total overnight investment costs of the project (per unit of capacity), data.frame with columns:

region

character, name of the region for which the parameter will be applied, NA for every region

year

integer, year for which the parameter will be applied, NA for every year

invcost

numeric, total overnight investment costs of the project (per unit of capacity), NA means default value (0)

start

the first year the technology can be installed, data.frame with columns:

region

character, name of the region for which the parameter will be applied, NA for every region

start

integer, the first year the technology can be installed, NA means all years of the model horizon

end

the last year the technology can be installed, data.frame with columns:

region

character, name of the region for which the parameter will be applied, NA means default value (-Inf)

end

integer, the last year the technology can be installed, NA means default value (Inf)

olife

operational life of the installed technology (in years), data.frame with columns:

region

character, name of the region for which the parameter will be applied, NA for every region

olife

integer, operational life of the installed technology (in years), NA means default value (1)

stock

predefined stock of the installed technology (in units of capacity), data.frame with columns:

region

character, name of the region for which the parameter will be applied, NA for every region

year

integer, year for which the stock will be specified, required, values between specified years will be interpolated.

stock

numeric, predefined existing of future capacity of the technology (in units of capacity), NA means default value (0)

optimizeRetirement

logical. Currently ignored.

upgrade.technology

character. Currently ignored.

fullYear

logical. Currently ignored.

timeframe

character name of timeframe level the technology is operating. By default, the lowest level of commodities in the technology is applied.

region

character vector of regions where the technology exist or can be installed.

misc

list with any miscellaneous information.

capacity

data.frame (not implemented!) Capacity parameters of the technology.