UTOPIA electricity load shape

A deterministic relative electricity-load shape by timeslice, for the three teaching calendars (replaces the vignette's former random load curve). utopia_profiles() / the vignette scale it by a region's annual demand and the timeslice shares to get energy per timeslice.

Usage

utopia_demand

Format

A data.frame with columns calendar (s4_h24/m12_h24/utopia_seasons), timeslice and load (relative, mean ~1).

See also

Examples

utopia_demand
#>           calendar timeslice   load
#> 1           s4_h24   WIN_h00 0.8400
#> 2           s4_h24   SPR_h00 0.6300
#> 3           s4_h24   SUM_h00 0.7700
#> 4           s4_h24   FAL_h00 0.6300
#> 5           s4_h24   WIN_h01 0.7800
#> 6           s4_h24   SPR_h01 0.5850
#> 7           s4_h24   SUM_h01 0.7150
#> 8           s4_h24   FAL_h01 0.5850
#> 9           s4_h24   WIN_h02 0.7440
#> 10          s4_h24   SPR_h02 0.5580
#> 11          s4_h24   SUM_h02 0.6820
#> 12          s4_h24   FAL_h02 0.5580
#> 13          s4_h24   WIN_h03 0.7200
#> 14          s4_h24   SPR_h03 0.5400
#> 15          s4_h24   SUM_h03 0.6600
#> 16          s4_h24   FAL_h03 0.5400
#> 17          s4_h24   WIN_h04 0.7440
#> 18          s4_h24   SPR_h04 0.5580
#> 19          s4_h24   SUM_h04 0.6820
#> 20          s4_h24   FAL_h04 0.5580
#> 21          s4_h24   WIN_h05 0.8160
#> 22          s4_h24   SPR_h05 0.6120
#> 23          s4_h24   SUM_h05 0.7480
#> 24          s4_h24   FAL_h05 0.6120
#> 25          s4_h24   WIN_h06 0.9600
#> 26          s4_h24   SPR_h06 0.7200
#> 27          s4_h24   SUM_h06 0.8800
#> 28          s4_h24   FAL_h06 0.7200
#> 29          s4_h24   WIN_h07 1.1400
#> 30          s4_h24   SPR_h07 0.8550
#> 31          s4_h24   SUM_h07 1.0450
#> 32          s4_h24   FAL_h07 0.8550
#> 33          s4_h24   WIN_h08 1.2600
#> 34          s4_h24   SPR_h08 0.9450
#> 35          s4_h24   SUM_h08 1.1550
#> 36          s4_h24   FAL_h08 0.9450
#> 37          s4_h24   WIN_h09 1.2960
#> 38          s4_h24   SPR_h09 0.9720
#> 39          s4_h24   SUM_h09 1.1880
#> 40          s4_h24   FAL_h09 0.9720
#> 41          s4_h24   WIN_h10 1.2840
#> 42          s4_h24   SPR_h10 0.9630
#> 43          s4_h24   SUM_h10 1.1770
#> 44          s4_h24   FAL_h10 0.9630
#> 45          s4_h24   WIN_h11 1.2600
#> 46          s4_h24   SPR_h11 0.9450
#> 47          s4_h24   SUM_h11 1.1550
#> 48          s4_h24   FAL_h11 0.9450
#> 49          s4_h24   WIN_h12 1.2600
#> 50          s4_h24   SPR_h12 0.9450
#> 51          s4_h24   SUM_h12 1.1550
#> 52          s4_h24   FAL_h12 0.9450
#> 53          s4_h24   WIN_h13 1.2480
#> 54          s4_h24   SPR_h13 0.9360
#> 55          s4_h24   SUM_h13 1.1440
#> 56          s4_h24   FAL_h13 0.9360
#> 57          s4_h24   WIN_h14 1.2360
#> 58          s4_h24   SPR_h14 0.9270
#> 59          s4_h24   SUM_h14 1.1330
#> 60          s4_h24   FAL_h14 0.9270
#> 61          s4_h24   WIN_h15 1.2600
#> 62          s4_h24   SPR_h15 0.9450
#> 63          s4_h24   SUM_h15 1.1550
#> 64          s4_h24   FAL_h15 0.9450
#> 65          s4_h24   WIN_h16 1.3200
#> 66          s4_h24   SPR_h16 0.9900
#> 67          s4_h24   SUM_h16 1.2100
#> 68          s4_h24   FAL_h16 0.9900
#> 69          s4_h24   WIN_h17 1.4400
#> 70          s4_h24   SPR_h17 1.0800
#> 71          s4_h24   SUM_h17 1.3200
#> 72          s4_h24   FAL_h17 1.0800
#> 73          s4_h24   WIN_h18 1.5600
#> 74          s4_h24   SPR_h18 1.1700
#> 75          s4_h24   SUM_h18 1.4300
#> 76          s4_h24   FAL_h18 1.1700
#> 77          s4_h24   WIN_h19 1.5840
#> 78          s4_h24   SPR_h19 1.1880
#> 79          s4_h24   SUM_h19 1.4520
#> 80          s4_h24   FAL_h19 1.1880
#> 81          s4_h24   WIN_h20 1.5000
#> 82          s4_h24   SPR_h20 1.1250
#> 83          s4_h24   SUM_h20 1.3750
#> 84          s4_h24   FAL_h20 1.1250
#> 85          s4_h24   WIN_h21 1.3200
#> 86          s4_h24   SPR_h21 0.9900
#> 87          s4_h24   SUM_h21 1.2100
#> 88          s4_h24   FAL_h21 0.9900
#> 89          s4_h24   WIN_h22 1.1400
#> 90          s4_h24   SPR_h22 0.8550
#> 91          s4_h24   SUM_h22 1.0450
#> 92          s4_h24   FAL_h22 0.8550
#> 93          s4_h24   WIN_h23 0.9600
#> 94          s4_h24   SPR_h23 0.7200
#> 95          s4_h24   SUM_h23 0.8800
#> 96          s4_h24   FAL_h23 0.7200
#> 97         m12_h24   m01_h00 0.8400
#> 98         m12_h24   m02_h00 0.8050
#> 99         m12_h24   m03_h00 0.7000
#> 100        m12_h24   m04_h00 0.6300
#> 101        m12_h24   m05_h00 0.6300
#> 102        m12_h24   m06_h00 0.7000
#> 103        m12_h24   m07_h00 0.7700
#> 104        m12_h24   m08_h00 0.7700
#> 105        m12_h24   m09_h00 0.6650
#> 106        m12_h24   m10_h00 0.6300
#> 107        m12_h24   m11_h00 0.7000
#> 108        m12_h24   m12_h00 0.8400
#> 109        m12_h24   m01_h01 0.7800
#> 110        m12_h24   m02_h01 0.7475
#> 111        m12_h24   m03_h01 0.6500
#> 112        m12_h24   m04_h01 0.5850
#> 113        m12_h24   m05_h01 0.5850
#> 114        m12_h24   m06_h01 0.6500
#> 115        m12_h24   m07_h01 0.7150
#> 116        m12_h24   m08_h01 0.7150
#> 117        m12_h24   m09_h01 0.6175
#> 118        m12_h24   m10_h01 0.5850
#> 119        m12_h24   m11_h01 0.6500
#> 120        m12_h24   m12_h01 0.7800
#> 121        m12_h24   m01_h02 0.7440
#> 122        m12_h24   m02_h02 0.7130
#> 123        m12_h24   m03_h02 0.6200
#> 124        m12_h24   m04_h02 0.5580
#> 125        m12_h24   m05_h02 0.5580
#> 126        m12_h24   m06_h02 0.6200
#> 127        m12_h24   m07_h02 0.6820
#> 128        m12_h24   m08_h02 0.6820
#> 129        m12_h24   m09_h02 0.5890
#> 130        m12_h24   m10_h02 0.5580
#> 131        m12_h24   m11_h02 0.6200
#> 132        m12_h24   m12_h02 0.7440
#> 133        m12_h24   m01_h03 0.7200
#> 134        m12_h24   m02_h03 0.6900
#> 135        m12_h24   m03_h03 0.6000
#> 136        m12_h24   m04_h03 0.5400
#> 137        m12_h24   m05_h03 0.5400
#> 138        m12_h24   m06_h03 0.6000
#> 139        m12_h24   m07_h03 0.6600
#> 140        m12_h24   m08_h03 0.6600
#> 141        m12_h24   m09_h03 0.5700
#> 142        m12_h24   m10_h03 0.5400
#> 143        m12_h24   m11_h03 0.6000
#> 144        m12_h24   m12_h03 0.7200
#> 145        m12_h24   m01_h04 0.7440
#> 146        m12_h24   m02_h04 0.7130
#> 147        m12_h24   m03_h04 0.6200
#> 148        m12_h24   m04_h04 0.5580
#> 149        m12_h24   m05_h04 0.5580
#> 150        m12_h24   m06_h04 0.6200
#> 151        m12_h24   m07_h04 0.6820
#> 152        m12_h24   m08_h04 0.6820
#> 153        m12_h24   m09_h04 0.5890
#> 154        m12_h24   m10_h04 0.5580
#> 155        m12_h24   m11_h04 0.6200
#> 156        m12_h24   m12_h04 0.7440
#> 157        m12_h24   m01_h05 0.8160
#> 158        m12_h24   m02_h05 0.7820
#> 159        m12_h24   m03_h05 0.6800
#> 160        m12_h24   m04_h05 0.6120
#> 161        m12_h24   m05_h05 0.6120
#> 162        m12_h24   m06_h05 0.6800
#> 163        m12_h24   m07_h05 0.7480
#> 164        m12_h24   m08_h05 0.7480
#> 165        m12_h24   m09_h05 0.6460
#> 166        m12_h24   m10_h05 0.6120
#> 167        m12_h24   m11_h05 0.6800
#> 168        m12_h24   m12_h05 0.8160
#> 169        m12_h24   m01_h06 0.9600
#> 170        m12_h24   m02_h06 0.9200
#> 171        m12_h24   m03_h06 0.8000
#> 172        m12_h24   m04_h06 0.7200
#> 173        m12_h24   m05_h06 0.7200
#> 174        m12_h24   m06_h06 0.8000
#> 175        m12_h24   m07_h06 0.8800
#> 176        m12_h24   m08_h06 0.8800
#> 177        m12_h24   m09_h06 0.7600
#> 178        m12_h24   m10_h06 0.7200
#> 179        m12_h24   m11_h06 0.8000
#> 180        m12_h24   m12_h06 0.9600
#> 181        m12_h24   m01_h07 1.1400
#> 182        m12_h24   m02_h07 1.0925
#> 183        m12_h24   m03_h07 0.9500
#> 184        m12_h24   m04_h07 0.8550
#> 185        m12_h24   m05_h07 0.8550
#> 186        m12_h24   m06_h07 0.9500
#> 187        m12_h24   m07_h07 1.0450
#> 188        m12_h24   m08_h07 1.0450
#> 189        m12_h24   m09_h07 0.9025
#> 190        m12_h24   m10_h07 0.8550
#> 191        m12_h24   m11_h07 0.9500
#> 192        m12_h24   m12_h07 1.1400
#> 193        m12_h24   m01_h08 1.2600
#> 194        m12_h24   m02_h08 1.2075
#> 195        m12_h24   m03_h08 1.0500
#> 196        m12_h24   m04_h08 0.9450
#> 197        m12_h24   m05_h08 0.9450
#> 198        m12_h24   m06_h08 1.0500
#> 199        m12_h24   m07_h08 1.1550
#> 200        m12_h24   m08_h08 1.1550
#> 201        m12_h24   m09_h08 0.9975
#> 202        m12_h24   m10_h08 0.9450
#> 203        m12_h24   m11_h08 1.0500
#> 204        m12_h24   m12_h08 1.2600
#> 205        m12_h24   m01_h09 1.2960
#> 206        m12_h24   m02_h09 1.2420
#> 207        m12_h24   m03_h09 1.0800
#> 208        m12_h24   m04_h09 0.9720
#> 209        m12_h24   m05_h09 0.9720
#> 210        m12_h24   m06_h09 1.0800
#> 211        m12_h24   m07_h09 1.1880
#> 212        m12_h24   m08_h09 1.1880
#> 213        m12_h24   m09_h09 1.0260
#> 214        m12_h24   m10_h09 0.9720
#> 215        m12_h24   m11_h09 1.0800
#> 216        m12_h24   m12_h09 1.2960
#> 217        m12_h24   m01_h10 1.2840
#> 218        m12_h24   m02_h10 1.2305
#> 219        m12_h24   m03_h10 1.0700
#> 220        m12_h24   m04_h10 0.9630
#> 221        m12_h24   m05_h10 0.9630
#> 222        m12_h24   m06_h10 1.0700
#> 223        m12_h24   m07_h10 1.1770
#> 224        m12_h24   m08_h10 1.1770
#> 225        m12_h24   m09_h10 1.0165
#> 226        m12_h24   m10_h10 0.9630
#> 227        m12_h24   m11_h10 1.0700
#> 228        m12_h24   m12_h10 1.2840
#> 229        m12_h24   m01_h11 1.2600
#> 230        m12_h24   m02_h11 1.2075
#> 231        m12_h24   m03_h11 1.0500
#> 232        m12_h24   m04_h11 0.9450
#> 233        m12_h24   m05_h11 0.9450
#> 234        m12_h24   m06_h11 1.0500
#> 235        m12_h24   m07_h11 1.1550
#> 236        m12_h24   m08_h11 1.1550
#> 237        m12_h24   m09_h11 0.9975
#> 238        m12_h24   m10_h11 0.9450
#> 239        m12_h24   m11_h11 1.0500
#> 240        m12_h24   m12_h11 1.2600
#> 241        m12_h24   m01_h12 1.2600
#> 242        m12_h24   m02_h12 1.2075
#> 243        m12_h24   m03_h12 1.0500
#> 244        m12_h24   m04_h12 0.9450
#> 245        m12_h24   m05_h12 0.9450
#> 246        m12_h24   m06_h12 1.0500
#> 247        m12_h24   m07_h12 1.1550
#> 248        m12_h24   m08_h12 1.1550
#> 249        m12_h24   m09_h12 0.9975
#> 250        m12_h24   m10_h12 0.9450
#> 251        m12_h24   m11_h12 1.0500
#> 252        m12_h24   m12_h12 1.2600
#> 253        m12_h24   m01_h13 1.2480
#> 254        m12_h24   m02_h13 1.1960
#> 255        m12_h24   m03_h13 1.0400
#> 256        m12_h24   m04_h13 0.9360
#> 257        m12_h24   m05_h13 0.9360
#> 258        m12_h24   m06_h13 1.0400
#> 259        m12_h24   m07_h13 1.1440
#> 260        m12_h24   m08_h13 1.1440
#> 261        m12_h24   m09_h13 0.9880
#> 262        m12_h24   m10_h13 0.9360
#> 263        m12_h24   m11_h13 1.0400
#> 264        m12_h24   m12_h13 1.2480
#> 265        m12_h24   m01_h14 1.2360
#> 266        m12_h24   m02_h14 1.1845
#> 267        m12_h24   m03_h14 1.0300
#> 268        m12_h24   m04_h14 0.9270
#> 269        m12_h24   m05_h14 0.9270
#> 270        m12_h24   m06_h14 1.0300
#> 271        m12_h24   m07_h14 1.1330
#> 272        m12_h24   m08_h14 1.1330
#> 273        m12_h24   m09_h14 0.9785
#> 274        m12_h24   m10_h14 0.9270
#> 275        m12_h24   m11_h14 1.0300
#> 276        m12_h24   m12_h14 1.2360
#> 277        m12_h24   m01_h15 1.2600
#> 278        m12_h24   m02_h15 1.2075
#> 279        m12_h24   m03_h15 1.0500
#> 280        m12_h24   m04_h15 0.9450
#> 281        m12_h24   m05_h15 0.9450
#> 282        m12_h24   m06_h15 1.0500
#> 283        m12_h24   m07_h15 1.1550
#> 284        m12_h24   m08_h15 1.1550
#> 285        m12_h24   m09_h15 0.9975
#> 286        m12_h24   m10_h15 0.9450
#> 287        m12_h24   m11_h15 1.0500
#> 288        m12_h24   m12_h15 1.2600
#> 289        m12_h24   m01_h16 1.3200
#> 290        m12_h24   m02_h16 1.2650
#> 291        m12_h24   m03_h16 1.1000
#> 292        m12_h24   m04_h16 0.9900
#> 293        m12_h24   m05_h16 0.9900
#> 294        m12_h24   m06_h16 1.1000
#> 295        m12_h24   m07_h16 1.2100
#> 296        m12_h24   m08_h16 1.2100
#> 297        m12_h24   m09_h16 1.0450
#> 298        m12_h24   m10_h16 0.9900
#> 299        m12_h24   m11_h16 1.1000
#> 300        m12_h24   m12_h16 1.3200
#> 301        m12_h24   m01_h17 1.4400
#> 302        m12_h24   m02_h17 1.3800
#> 303        m12_h24   m03_h17 1.2000
#> 304        m12_h24   m04_h17 1.0800
#> 305        m12_h24   m05_h17 1.0800
#> 306        m12_h24   m06_h17 1.2000
#> 307        m12_h24   m07_h17 1.3200
#> 308        m12_h24   m08_h17 1.3200
#> 309        m12_h24   m09_h17 1.1400
#> 310        m12_h24   m10_h17 1.0800
#> 311        m12_h24   m11_h17 1.2000
#> 312        m12_h24   m12_h17 1.4400
#> 313        m12_h24   m01_h18 1.5600
#> 314        m12_h24   m02_h18 1.4950
#> 315        m12_h24   m03_h18 1.3000
#> 316        m12_h24   m04_h18 1.1700
#> 317        m12_h24   m05_h18 1.1700
#> 318        m12_h24   m06_h18 1.3000
#> 319        m12_h24   m07_h18 1.4300
#> 320        m12_h24   m08_h18 1.4300
#> 321        m12_h24   m09_h18 1.2350
#> 322        m12_h24   m10_h18 1.1700
#> 323        m12_h24   m11_h18 1.3000
#> 324        m12_h24   m12_h18 1.5600
#> 325        m12_h24   m01_h19 1.5840
#> 326        m12_h24   m02_h19 1.5180
#> 327        m12_h24   m03_h19 1.3200
#> 328        m12_h24   m04_h19 1.1880
#> 329        m12_h24   m05_h19 1.1880
#> 330        m12_h24   m06_h19 1.3200
#> 331        m12_h24   m07_h19 1.4520
#> 332        m12_h24   m08_h19 1.4520
#> 333        m12_h24   m09_h19 1.2540
#> 334        m12_h24   m10_h19 1.1880
#> 335        m12_h24   m11_h19 1.3200
#> 336        m12_h24   m12_h19 1.5840
#> 337        m12_h24   m01_h20 1.5000
#> 338        m12_h24   m02_h20 1.4375
#> 339        m12_h24   m03_h20 1.2500
#> 340        m12_h24   m04_h20 1.1250
#> 341        m12_h24   m05_h20 1.1250
#> 342        m12_h24   m06_h20 1.2500
#> 343        m12_h24   m07_h20 1.3750
#> 344        m12_h24   m08_h20 1.3750
#> 345        m12_h24   m09_h20 1.1875
#> 346        m12_h24   m10_h20 1.1250
#> 347        m12_h24   m11_h20 1.2500
#> 348        m12_h24   m12_h20 1.5000
#> 349        m12_h24   m01_h21 1.3200
#> 350        m12_h24   m02_h21 1.2650
#> 351        m12_h24   m03_h21 1.1000
#> 352        m12_h24   m04_h21 0.9900
#> 353        m12_h24   m05_h21 0.9900
#> 354        m12_h24   m06_h21 1.1000
#> 355        m12_h24   m07_h21 1.2100
#> 356        m12_h24   m08_h21 1.2100
#> 357        m12_h24   m09_h21 1.0450
#> 358        m12_h24   m10_h21 0.9900
#> 359        m12_h24   m11_h21 1.1000
#> 360        m12_h24   m12_h21 1.3200
#> 361        m12_h24   m01_h22 1.1400
#> 362        m12_h24   m02_h22 1.0925
#> 363        m12_h24   m03_h22 0.9500
#> 364        m12_h24   m04_h22 0.8550
#> 365        m12_h24   m05_h22 0.8550
#> 366        m12_h24   m06_h22 0.9500
#> 367        m12_h24   m07_h22 1.0450
#> 368        m12_h24   m08_h22 1.0450
#> 369        m12_h24   m09_h22 0.9025
#> 370        m12_h24   m10_h22 0.8550
#> 371        m12_h24   m11_h22 0.9500
#> 372        m12_h24   m12_h22 1.1400
#> 373        m12_h24   m01_h23 0.9600
#> 374        m12_h24   m02_h23 0.9200
#> 375        m12_h24   m03_h23 0.8000
#> 376        m12_h24   m04_h23 0.7200
#> 377        m12_h24   m05_h23 0.7200
#> 378        m12_h24   m06_h23 0.8000
#> 379        m12_h24   m07_h23 0.8800
#> 380        m12_h24   m08_h23 0.8800
#> 381        m12_h24   m09_h23 0.7600
#> 382        m12_h24   m10_h23 0.7200
#> 383        m12_h24   m11_h23 0.8000
#> 384        m12_h24   m12_h23 0.9600
#> 385 utopia_seasons   WIN_DAY 1.2000
#> 386 utopia_seasons   SPR_DAY 0.9000
#> 387 utopia_seasons   SUM_DAY 1.1000
#> 388 utopia_seasons   FAL_DAY 0.9000
#> 389 utopia_seasons   WIN_NGT 0.7200
#> 390 utopia_seasons   SPR_NGT 0.5400
#> 391 utopia_seasons   SUM_NGT 0.6600
#> 392 utopia_seasons   FAL_NGT 0.5400
#> 393 utopia_seasons    WIN_PK 1.6200
#> 394 utopia_seasons    SPR_PK 1.2150
#> 395 utopia_seasons    SUM_PK 1.4850
#> 396 utopia_seasons    FAL_PK 1.2150