Convert year-days to YDAY set 'dNNN'

Convert year-days to YDAY set 'dNNN'

Usage

yday2YDAY(x, width = 3, prefix = "d", flag = "0")

Arguments

x

integer vector, year-days (for example, 1-365 for annual data)

width

integer, width of the output string, default is 3

prefix

character, prefix to add to the name, default is 'd'

flag

character, flag to add to the name, default is '0'

Value

character vector of the same length as x with formatted year-days to be used in the YDAY set.

Examples

yday2YDAY(1:365)
#>   [1] "d001" "d002" "d003" "d004" "d005" "d006" "d007" "d008" "d009" "d010"
#>  [11] "d011" "d012" "d013" "d014" "d015" "d016" "d017" "d018" "d019" "d020"
#>  [21] "d021" "d022" "d023" "d024" "d025" "d026" "d027" "d028" "d029" "d030"
#>  [31] "d031" "d032" "d033" "d034" "d035" "d036" "d037" "d038" "d039" "d040"
#>  [41] "d041" "d042" "d043" "d044" "d045" "d046" "d047" "d048" "d049" "d050"
#>  [51] "d051" "d052" "d053" "d054" "d055" "d056" "d057" "d058" "d059" "d060"
#>  [61] "d061" "d062" "d063" "d064" "d065" "d066" "d067" "d068" "d069" "d070"
#>  [71] "d071" "d072" "d073" "d074" "d075" "d076" "d077" "d078" "d079" "d080"
#>  [81] "d081" "d082" "d083" "d084" "d085" "d086" "d087" "d088" "d089" "d090"
#>  [91] "d091" "d092" "d093" "d094" "d095" "d096" "d097" "d098" "d099" "d100"
#> [101] "d101" "d102" "d103" "d104" "d105" "d106" "d107" "d108" "d109" "d110"
#> [111] "d111" "d112" "d113" "d114" "d115" "d116" "d117" "d118" "d119" "d120"
#> [121] "d121" "d122" "d123" "d124" "d125" "d126" "d127" "d128" "d129" "d130"
#> [131] "d131" "d132" "d133" "d134" "d135" "d136" "d137" "d138" "d139" "d140"
#> [141] "d141" "d142" "d143" "d144" "d145" "d146" "d147" "d148" "d149" "d150"
#> [151] "d151" "d152" "d153" "d154" "d155" "d156" "d157" "d158" "d159" "d160"
#> [161] "d161" "d162" "d163" "d164" "d165" "d166" "d167" "d168" "d169" "d170"
#> [171] "d171" "d172" "d173" "d174" "d175" "d176" "d177" "d178" "d179" "d180"
#> [181] "d181" "d182" "d183" "d184" "d185" "d186" "d187" "d188" "d189" "d190"
#> [191] "d191" "d192" "d193" "d194" "d195" "d196" "d197" "d198" "d199" "d200"
#> [201] "d201" "d202" "d203" "d204" "d205" "d206" "d207" "d208" "d209" "d210"
#> [211] "d211" "d212" "d213" "d214" "d215" "d216" "d217" "d218" "d219" "d220"
#> [221] "d221" "d222" "d223" "d224" "d225" "d226" "d227" "d228" "d229" "d230"
#> [231] "d231" "d232" "d233" "d234" "d235" "d236" "d237" "d238" "d239" "d240"
#> [241] "d241" "d242" "d243" "d244" "d245" "d246" "d247" "d248" "d249" "d250"
#> [251] "d251" "d252" "d253" "d254" "d255" "d256" "d257" "d258" "d259" "d260"
#> [261] "d261" "d262" "d263" "d264" "d265" "d266" "d267" "d268" "d269" "d270"
#> [271] "d271" "d272" "d273" "d274" "d275" "d276" "d277" "d278" "d279" "d280"
#> [281] "d281" "d282" "d283" "d284" "d285" "d286" "d287" "d288" "d289" "d290"
#> [291] "d291" "d292" "d293" "d294" "d295" "d296" "d297" "d298" "d299" "d300"
#> [301] "d301" "d302" "d303" "d304" "d305" "d306" "d307" "d308" "d309" "d310"
#> [311] "d311" "d312" "d313" "d314" "d315" "d316" "d317" "d318" "d319" "d320"
#> [321] "d321" "d322" "d323" "d324" "d325" "d326" "d327" "d328" "d329" "d330"
#> [331] "d331" "d332" "d333" "d334" "d335" "d336" "d337" "d338" "d339" "d340"
#> [341] "d341" "d342" "d343" "d344" "d345" "d346" "d347" "d348" "d349" "d350"
#> [351] "d351" "d352" "d353" "d354" "d355" "d356" "d357" "d358" "d359" "d360"
#> [361] "d361" "d362" "d363" "d364" "d365"