An S4 class to represent a custom constraint.

Class constraint is used to define custom constraints in the optimization problem.

Details

Slots

name

character. Name of the constraint object, used in sets.

desc

character. Description of the constraint.

eq

character. Type of the relation ('==' default, '<=', '>=').

for.each

list. List with sets for combination of which the constraint is created.

rhs

data.frame. Named list or data frame with numeric values for each constraint. The dimensions of the data frame should match the dimensions of the sets in the for.each slot.

defVal

numeric. The default value for the rhs. It is recommended to set the default value for the rhs of every constraint to avoid unexpected behavior. If not specified, the default value is 0, and the warning is issued.

interpolation

character. Interpolation rule for the constraint. Recognized values, any combination of "back", "inter", "forth", e.g., "back.inter" or "forth.inter", indicating the direction of interpolation. The default value is "inter", meaning that the interpolation is done for years between the specified values. The "back" and "forth" values induce backward and forward interpolation of the rhs values, respectively.

lhs

list. List of summands for the left-hand-side of the equation. This slot is created automatically from all named of unnamed lists passed to the newConstraint function, except for the named arguments.

misc

list. Any additional information or parameters to store in the constraint object.

See also