Performes search for available data in _scenario_ object.

Performes search for available data in _scenario_ object.

findData(
  scen,
  dataType = c("parameters", "variables"),
  setsNames_ = NULL,
  valueColumn = TRUE,
  allSets = TRUE,
  ignore.case = FALSE,
  dropEmpty = TRUE,
  dfDim = TRUE,
  dfNames = TRUE,
  asMatrix = FALSE
)

Arguments

scen

object _scenario_ with model solution.

dataType

type of data to lok for (currently only "parameters" and "variables").

setsNames_

regular expression pattern for names of sets which will be included in search.

valueColumn

logical, if TRUE will return variables and parameters with 'value' column (to filter sets and mappings).

allSets

logical, if TRUE _and_ operator should be used in search the sets, _or_ will be used if FALSE.

ignore.case

grepl parameter for matchin names.

dropEmpty

logical, if TRUE drops parameters and variables with zero length.

dfDim

logical, if TRUE returns dimension _dim_.

dfNames

logical, when TRUE returns names of the data frame column.

asMatrix

return results as a matric (not implemented).

Value

list with variables and parameters name, each includes _dim_ and _names_ character vectors.