Calculate levelized costs

levcost is a method for class technology to calculate levelized costs of production of commodity

# S4 method for repository
levcost(obj, ...)

Arguments

tech
object of class technology

start_year
numeric, the year of investments

discount
numeric, the discount rate

price
data frame with input commodity prices, has three mandatory fields:

comm character, names of commodities

price numeric, prices of the commodity

year numeric, years

solver
character, name of solver software "GAMS" or "GLPK"

tmp.dir
a dirrectory for temporary files to run the model, by default '/solwork' in the project directory.

tmp.del
a logical scalar indicating if solver temporary files should be deleted.