Create (equality or inequality) constraint object

Create (equality or inequality) constraint object

Usage

newConstraint(
  name,
  desc = "",
  ...,
  eq = "==",
  rhs = data.frame(),
  for.each = NULL,
  defVal = NULL,
  interpolation = "inter",
  replace_zerros = 1e-20,
  arg = NULL
)

Arguments

name

Name of the constrain object (will be used in GAMS or GLPK as an element in sets)

...

Left-hand side (LHS) terms of the statement - list objects with

eq

Type of the relation ('==' default, '<=', '>=')

rhs

a numeric value, list or data frame with sets and numeric values for each constraint. Warning: zero values will be replaced with `1e-20` to avoid ignoring them by the current interpolation algorithms.

for.each

list with sets for which constraint will be created.

defVal

the default value for the `rhs`.

interpolation

interpolation rule for the constraint. Recognized values, any combination of "back", "inter", "forth", indicating the direction of interpolation. The default value is "inter".

arg

tbc

Value

Object of class `constraint`.