UTOPIA electricity load shape

A deterministic relative electricity-load shape by slice, for both teaching calendars (replaces the vignette's former random load curve). utopia_profiles() / the vignette scale it by a region's annual demand and the slice shares to get energy per slice.

Usage

utopia_demand

Format

A data.frame with columns calendar (utopia_m12h24/utopia_seasons), slice and load (relative, mean ~1).

See also

Examples

utopia_demand
#>           calendar   slice   load
#> 1     utopia_s4h24 WIN_h00 0.8400
#> 2     utopia_s4h24 SPR_h00 0.6300
#> 3     utopia_s4h24 SUM_h00 0.7700
#> 4     utopia_s4h24 AUT_h00 0.6300
#> 5     utopia_s4h24 WIN_h01 0.7800
#> 6     utopia_s4h24 SPR_h01 0.5850
#> 7     utopia_s4h24 SUM_h01 0.7150
#> 8     utopia_s4h24 AUT_h01 0.5850
#> 9     utopia_s4h24 WIN_h02 0.7440
#> 10    utopia_s4h24 SPR_h02 0.5580
#> 11    utopia_s4h24 SUM_h02 0.6820
#> 12    utopia_s4h24 AUT_h02 0.5580
#> 13    utopia_s4h24 WIN_h03 0.7200
#> 14    utopia_s4h24 SPR_h03 0.5400
#> 15    utopia_s4h24 SUM_h03 0.6600
#> 16    utopia_s4h24 AUT_h03 0.5400
#> 17    utopia_s4h24 WIN_h04 0.7440
#> 18    utopia_s4h24 SPR_h04 0.5580
#> 19    utopia_s4h24 SUM_h04 0.6820
#> 20    utopia_s4h24 AUT_h04 0.5580
#> 21    utopia_s4h24 WIN_h05 0.8160
#> 22    utopia_s4h24 SPR_h05 0.6120
#> 23    utopia_s4h24 SUM_h05 0.7480
#> 24    utopia_s4h24 AUT_h05 0.6120
#> 25    utopia_s4h24 WIN_h06 0.9600
#> 26    utopia_s4h24 SPR_h06 0.7200
#> 27    utopia_s4h24 SUM_h06 0.8800
#> 28    utopia_s4h24 AUT_h06 0.7200
#> 29    utopia_s4h24 WIN_h07 1.1400
#> 30    utopia_s4h24 SPR_h07 0.8550
#> 31    utopia_s4h24 SUM_h07 1.0450
#> 32    utopia_s4h24 AUT_h07 0.8550
#> 33    utopia_s4h24 WIN_h08 1.2600
#> 34    utopia_s4h24 SPR_h08 0.9450
#> 35    utopia_s4h24 SUM_h08 1.1550
#> 36    utopia_s4h24 AUT_h08 0.9450
#> 37    utopia_s4h24 WIN_h09 1.2960
#> 38    utopia_s4h24 SPR_h09 0.9720
#> 39    utopia_s4h24 SUM_h09 1.1880
#> 40    utopia_s4h24 AUT_h09 0.9720
#> 41    utopia_s4h24 WIN_h10 1.2840
#> 42    utopia_s4h24 SPR_h10 0.9630
#> 43    utopia_s4h24 SUM_h10 1.1770
#> 44    utopia_s4h24 AUT_h10 0.9630
#> 45    utopia_s4h24 WIN_h11 1.2600
#> 46    utopia_s4h24 SPR_h11 0.9450
#> 47    utopia_s4h24 SUM_h11 1.1550
#> 48    utopia_s4h24 AUT_h11 0.9450
#> 49    utopia_s4h24 WIN_h12 1.2600
#> 50    utopia_s4h24 SPR_h12 0.9450
#> 51    utopia_s4h24 SUM_h12 1.1550
#> 52    utopia_s4h24 AUT_h12 0.9450
#> 53    utopia_s4h24 WIN_h13 1.2480
#> 54    utopia_s4h24 SPR_h13 0.9360
#> 55    utopia_s4h24 SUM_h13 1.1440
#> 56    utopia_s4h24 AUT_h13 0.9360
#> 57    utopia_s4h24 WIN_h14 1.2360
#> 58    utopia_s4h24 SPR_h14 0.9270
#> 59    utopia_s4h24 SUM_h14 1.1330
#> 60    utopia_s4h24 AUT_h14 0.9270
#> 61    utopia_s4h24 WIN_h15 1.2600
#> 62    utopia_s4h24 SPR_h15 0.9450
#> 63    utopia_s4h24 SUM_h15 1.1550
#> 64    utopia_s4h24 AUT_h15 0.9450
#> 65    utopia_s4h24 WIN_h16 1.3200
#> 66    utopia_s4h24 SPR_h16 0.9900
#> 67    utopia_s4h24 SUM_h16 1.2100
#> 68    utopia_s4h24 AUT_h16 0.9900
#> 69    utopia_s4h24 WIN_h17 1.4400
#> 70    utopia_s4h24 SPR_h17 1.0800
#> 71    utopia_s4h24 SUM_h17 1.3200
#> 72    utopia_s4h24 AUT_h17 1.0800
#> 73    utopia_s4h24 WIN_h18 1.5600
#> 74    utopia_s4h24 SPR_h18 1.1700
#> 75    utopia_s4h24 SUM_h18 1.4300
#> 76    utopia_s4h24 AUT_h18 1.1700
#> 77    utopia_s4h24 WIN_h19 1.5840
#> 78    utopia_s4h24 SPR_h19 1.1880
#> 79    utopia_s4h24 SUM_h19 1.4520
#> 80    utopia_s4h24 AUT_h19 1.1880
#> 81    utopia_s4h24 WIN_h20 1.5000
#> 82    utopia_s4h24 SPR_h20 1.1250
#> 83    utopia_s4h24 SUM_h20 1.3750
#> 84    utopia_s4h24 AUT_h20 1.1250
#> 85    utopia_s4h24 WIN_h21 1.3200
#> 86    utopia_s4h24 SPR_h21 0.9900
#> 87    utopia_s4h24 SUM_h21 1.2100
#> 88    utopia_s4h24 AUT_h21 0.9900
#> 89    utopia_s4h24 WIN_h22 1.1400
#> 90    utopia_s4h24 SPR_h22 0.8550
#> 91    utopia_s4h24 SUM_h22 1.0450
#> 92    utopia_s4h24 AUT_h22 0.8550
#> 93    utopia_s4h24 WIN_h23 0.9600
#> 94    utopia_s4h24 SPR_h23 0.7200
#> 95    utopia_s4h24 SUM_h23 0.8800
#> 96    utopia_s4h24 AUT_h23 0.7200
#> 97   utopia_m12h24 m01_h00 0.8400
#> 98   utopia_m12h24 m02_h00 0.8050
#> 99   utopia_m12h24 m03_h00 0.7000
#> 100  utopia_m12h24 m04_h00 0.6300
#> 101  utopia_m12h24 m05_h00 0.6300
#> 102  utopia_m12h24 m06_h00 0.7000
#> 103  utopia_m12h24 m07_h00 0.7700
#> 104  utopia_m12h24 m08_h00 0.7700
#> 105  utopia_m12h24 m09_h00 0.6650
#> 106  utopia_m12h24 m10_h00 0.6300
#> 107  utopia_m12h24 m11_h00 0.7000
#> 108  utopia_m12h24 m12_h00 0.8400
#> 109  utopia_m12h24 m01_h01 0.7800
#> 110  utopia_m12h24 m02_h01 0.7475
#> 111  utopia_m12h24 m03_h01 0.6500
#> 112  utopia_m12h24 m04_h01 0.5850
#> 113  utopia_m12h24 m05_h01 0.5850
#> 114  utopia_m12h24 m06_h01 0.6500
#> 115  utopia_m12h24 m07_h01 0.7150
#> 116  utopia_m12h24 m08_h01 0.7150
#> 117  utopia_m12h24 m09_h01 0.6175
#> 118  utopia_m12h24 m10_h01 0.5850
#> 119  utopia_m12h24 m11_h01 0.6500
#> 120  utopia_m12h24 m12_h01 0.7800
#> 121  utopia_m12h24 m01_h02 0.7440
#> 122  utopia_m12h24 m02_h02 0.7130
#> 123  utopia_m12h24 m03_h02 0.6200
#> 124  utopia_m12h24 m04_h02 0.5580
#> 125  utopia_m12h24 m05_h02 0.5580
#> 126  utopia_m12h24 m06_h02 0.6200
#> 127  utopia_m12h24 m07_h02 0.6820
#> 128  utopia_m12h24 m08_h02 0.6820
#> 129  utopia_m12h24 m09_h02 0.5890
#> 130  utopia_m12h24 m10_h02 0.5580
#> 131  utopia_m12h24 m11_h02 0.6200
#> 132  utopia_m12h24 m12_h02 0.7440
#> 133  utopia_m12h24 m01_h03 0.7200
#> 134  utopia_m12h24 m02_h03 0.6900
#> 135  utopia_m12h24 m03_h03 0.6000
#> 136  utopia_m12h24 m04_h03 0.5400
#> 137  utopia_m12h24 m05_h03 0.5400
#> 138  utopia_m12h24 m06_h03 0.6000
#> 139  utopia_m12h24 m07_h03 0.6600
#> 140  utopia_m12h24 m08_h03 0.6600
#> 141  utopia_m12h24 m09_h03 0.5700
#> 142  utopia_m12h24 m10_h03 0.5400
#> 143  utopia_m12h24 m11_h03 0.6000
#> 144  utopia_m12h24 m12_h03 0.7200
#> 145  utopia_m12h24 m01_h04 0.7440
#> 146  utopia_m12h24 m02_h04 0.7130
#> 147  utopia_m12h24 m03_h04 0.6200
#> 148  utopia_m12h24 m04_h04 0.5580
#> 149  utopia_m12h24 m05_h04 0.5580
#> 150  utopia_m12h24 m06_h04 0.6200
#> 151  utopia_m12h24 m07_h04 0.6820
#> 152  utopia_m12h24 m08_h04 0.6820
#> 153  utopia_m12h24 m09_h04 0.5890
#> 154  utopia_m12h24 m10_h04 0.5580
#> 155  utopia_m12h24 m11_h04 0.6200
#> 156  utopia_m12h24 m12_h04 0.7440
#> 157  utopia_m12h24 m01_h05 0.8160
#> 158  utopia_m12h24 m02_h05 0.7820
#> 159  utopia_m12h24 m03_h05 0.6800
#> 160  utopia_m12h24 m04_h05 0.6120
#> 161  utopia_m12h24 m05_h05 0.6120
#> 162  utopia_m12h24 m06_h05 0.6800
#> 163  utopia_m12h24 m07_h05 0.7480
#> 164  utopia_m12h24 m08_h05 0.7480
#> 165  utopia_m12h24 m09_h05 0.6460
#> 166  utopia_m12h24 m10_h05 0.6120
#> 167  utopia_m12h24 m11_h05 0.6800
#> 168  utopia_m12h24 m12_h05 0.8160
#> 169  utopia_m12h24 m01_h06 0.9600
#> 170  utopia_m12h24 m02_h06 0.9200
#> 171  utopia_m12h24 m03_h06 0.8000
#> 172  utopia_m12h24 m04_h06 0.7200
#> 173  utopia_m12h24 m05_h06 0.7200
#> 174  utopia_m12h24 m06_h06 0.8000
#> 175  utopia_m12h24 m07_h06 0.8800
#> 176  utopia_m12h24 m08_h06 0.8800
#> 177  utopia_m12h24 m09_h06 0.7600
#> 178  utopia_m12h24 m10_h06 0.7200
#> 179  utopia_m12h24 m11_h06 0.8000
#> 180  utopia_m12h24 m12_h06 0.9600
#> 181  utopia_m12h24 m01_h07 1.1400
#> 182  utopia_m12h24 m02_h07 1.0925
#> 183  utopia_m12h24 m03_h07 0.9500
#> 184  utopia_m12h24 m04_h07 0.8550
#> 185  utopia_m12h24 m05_h07 0.8550
#> 186  utopia_m12h24 m06_h07 0.9500
#> 187  utopia_m12h24 m07_h07 1.0450
#> 188  utopia_m12h24 m08_h07 1.0450
#> 189  utopia_m12h24 m09_h07 0.9025
#> 190  utopia_m12h24 m10_h07 0.8550
#> 191  utopia_m12h24 m11_h07 0.9500
#> 192  utopia_m12h24 m12_h07 1.1400
#> 193  utopia_m12h24 m01_h08 1.2600
#> 194  utopia_m12h24 m02_h08 1.2075
#> 195  utopia_m12h24 m03_h08 1.0500
#> 196  utopia_m12h24 m04_h08 0.9450
#> 197  utopia_m12h24 m05_h08 0.9450
#> 198  utopia_m12h24 m06_h08 1.0500
#> 199  utopia_m12h24 m07_h08 1.1550
#> 200  utopia_m12h24 m08_h08 1.1550
#> 201  utopia_m12h24 m09_h08 0.9975
#> 202  utopia_m12h24 m10_h08 0.9450
#> 203  utopia_m12h24 m11_h08 1.0500
#> 204  utopia_m12h24 m12_h08 1.2600
#> 205  utopia_m12h24 m01_h09 1.2960
#> 206  utopia_m12h24 m02_h09 1.2420
#> 207  utopia_m12h24 m03_h09 1.0800
#> 208  utopia_m12h24 m04_h09 0.9720
#> 209  utopia_m12h24 m05_h09 0.9720
#> 210  utopia_m12h24 m06_h09 1.0800
#> 211  utopia_m12h24 m07_h09 1.1880
#> 212  utopia_m12h24 m08_h09 1.1880
#> 213  utopia_m12h24 m09_h09 1.0260
#> 214  utopia_m12h24 m10_h09 0.9720
#> 215  utopia_m12h24 m11_h09 1.0800
#> 216  utopia_m12h24 m12_h09 1.2960
#> 217  utopia_m12h24 m01_h10 1.2840
#> 218  utopia_m12h24 m02_h10 1.2305
#> 219  utopia_m12h24 m03_h10 1.0700
#> 220  utopia_m12h24 m04_h10 0.9630
#> 221  utopia_m12h24 m05_h10 0.9630
#> 222  utopia_m12h24 m06_h10 1.0700
#> 223  utopia_m12h24 m07_h10 1.1770
#> 224  utopia_m12h24 m08_h10 1.1770
#> 225  utopia_m12h24 m09_h10 1.0165
#> 226  utopia_m12h24 m10_h10 0.9630
#> 227  utopia_m12h24 m11_h10 1.0700
#> 228  utopia_m12h24 m12_h10 1.2840
#> 229  utopia_m12h24 m01_h11 1.2600
#> 230  utopia_m12h24 m02_h11 1.2075
#> 231  utopia_m12h24 m03_h11 1.0500
#> 232  utopia_m12h24 m04_h11 0.9450
#> 233  utopia_m12h24 m05_h11 0.9450
#> 234  utopia_m12h24 m06_h11 1.0500
#> 235  utopia_m12h24 m07_h11 1.1550
#> 236  utopia_m12h24 m08_h11 1.1550
#> 237  utopia_m12h24 m09_h11 0.9975
#> 238  utopia_m12h24 m10_h11 0.9450
#> 239  utopia_m12h24 m11_h11 1.0500
#> 240  utopia_m12h24 m12_h11 1.2600
#> 241  utopia_m12h24 m01_h12 1.2600
#> 242  utopia_m12h24 m02_h12 1.2075
#> 243  utopia_m12h24 m03_h12 1.0500
#> 244  utopia_m12h24 m04_h12 0.9450
#> 245  utopia_m12h24 m05_h12 0.9450
#> 246  utopia_m12h24 m06_h12 1.0500
#> 247  utopia_m12h24 m07_h12 1.1550
#> 248  utopia_m12h24 m08_h12 1.1550
#> 249  utopia_m12h24 m09_h12 0.9975
#> 250  utopia_m12h24 m10_h12 0.9450
#> 251  utopia_m12h24 m11_h12 1.0500
#> 252  utopia_m12h24 m12_h12 1.2600
#> 253  utopia_m12h24 m01_h13 1.2480
#> 254  utopia_m12h24 m02_h13 1.1960
#> 255  utopia_m12h24 m03_h13 1.0400
#> 256  utopia_m12h24 m04_h13 0.9360
#> 257  utopia_m12h24 m05_h13 0.9360
#> 258  utopia_m12h24 m06_h13 1.0400
#> 259  utopia_m12h24 m07_h13 1.1440
#> 260  utopia_m12h24 m08_h13 1.1440
#> 261  utopia_m12h24 m09_h13 0.9880
#> 262  utopia_m12h24 m10_h13 0.9360
#> 263  utopia_m12h24 m11_h13 1.0400
#> 264  utopia_m12h24 m12_h13 1.2480
#> 265  utopia_m12h24 m01_h14 1.2360
#> 266  utopia_m12h24 m02_h14 1.1845
#> 267  utopia_m12h24 m03_h14 1.0300
#> 268  utopia_m12h24 m04_h14 0.9270
#> 269  utopia_m12h24 m05_h14 0.9270
#> 270  utopia_m12h24 m06_h14 1.0300
#> 271  utopia_m12h24 m07_h14 1.1330
#> 272  utopia_m12h24 m08_h14 1.1330
#> 273  utopia_m12h24 m09_h14 0.9785
#> 274  utopia_m12h24 m10_h14 0.9270
#> 275  utopia_m12h24 m11_h14 1.0300
#> 276  utopia_m12h24 m12_h14 1.2360
#> 277  utopia_m12h24 m01_h15 1.2600
#> 278  utopia_m12h24 m02_h15 1.2075
#> 279  utopia_m12h24 m03_h15 1.0500
#> 280  utopia_m12h24 m04_h15 0.9450
#> 281  utopia_m12h24 m05_h15 0.9450
#> 282  utopia_m12h24 m06_h15 1.0500
#> 283  utopia_m12h24 m07_h15 1.1550
#> 284  utopia_m12h24 m08_h15 1.1550
#> 285  utopia_m12h24 m09_h15 0.9975
#> 286  utopia_m12h24 m10_h15 0.9450
#> 287  utopia_m12h24 m11_h15 1.0500
#> 288  utopia_m12h24 m12_h15 1.2600
#> 289  utopia_m12h24 m01_h16 1.3200
#> 290  utopia_m12h24 m02_h16 1.2650
#> 291  utopia_m12h24 m03_h16 1.1000
#> 292  utopia_m12h24 m04_h16 0.9900
#> 293  utopia_m12h24 m05_h16 0.9900
#> 294  utopia_m12h24 m06_h16 1.1000
#> 295  utopia_m12h24 m07_h16 1.2100
#> 296  utopia_m12h24 m08_h16 1.2100
#> 297  utopia_m12h24 m09_h16 1.0450
#> 298  utopia_m12h24 m10_h16 0.9900
#> 299  utopia_m12h24 m11_h16 1.1000
#> 300  utopia_m12h24 m12_h16 1.3200
#> 301  utopia_m12h24 m01_h17 1.4400
#> 302  utopia_m12h24 m02_h17 1.3800
#> 303  utopia_m12h24 m03_h17 1.2000
#> 304  utopia_m12h24 m04_h17 1.0800
#> 305  utopia_m12h24 m05_h17 1.0800
#> 306  utopia_m12h24 m06_h17 1.2000
#> 307  utopia_m12h24 m07_h17 1.3200
#> 308  utopia_m12h24 m08_h17 1.3200
#> 309  utopia_m12h24 m09_h17 1.1400
#> 310  utopia_m12h24 m10_h17 1.0800
#> 311  utopia_m12h24 m11_h17 1.2000
#> 312  utopia_m12h24 m12_h17 1.4400
#> 313  utopia_m12h24 m01_h18 1.5600
#> 314  utopia_m12h24 m02_h18 1.4950
#> 315  utopia_m12h24 m03_h18 1.3000
#> 316  utopia_m12h24 m04_h18 1.1700
#> 317  utopia_m12h24 m05_h18 1.1700
#> 318  utopia_m12h24 m06_h18 1.3000
#> 319  utopia_m12h24 m07_h18 1.4300
#> 320  utopia_m12h24 m08_h18 1.4300
#> 321  utopia_m12h24 m09_h18 1.2350
#> 322  utopia_m12h24 m10_h18 1.1700
#> 323  utopia_m12h24 m11_h18 1.3000
#> 324  utopia_m12h24 m12_h18 1.5600
#> 325  utopia_m12h24 m01_h19 1.5840
#> 326  utopia_m12h24 m02_h19 1.5180
#> 327  utopia_m12h24 m03_h19 1.3200
#> 328  utopia_m12h24 m04_h19 1.1880
#> 329  utopia_m12h24 m05_h19 1.1880
#> 330  utopia_m12h24 m06_h19 1.3200
#> 331  utopia_m12h24 m07_h19 1.4520
#> 332  utopia_m12h24 m08_h19 1.4520
#> 333  utopia_m12h24 m09_h19 1.2540
#> 334  utopia_m12h24 m10_h19 1.1880
#> 335  utopia_m12h24 m11_h19 1.3200
#> 336  utopia_m12h24 m12_h19 1.5840
#> 337  utopia_m12h24 m01_h20 1.5000
#> 338  utopia_m12h24 m02_h20 1.4375
#> 339  utopia_m12h24 m03_h20 1.2500
#> 340  utopia_m12h24 m04_h20 1.1250
#> 341  utopia_m12h24 m05_h20 1.1250
#> 342  utopia_m12h24 m06_h20 1.2500
#> 343  utopia_m12h24 m07_h20 1.3750
#> 344  utopia_m12h24 m08_h20 1.3750
#> 345  utopia_m12h24 m09_h20 1.1875
#> 346  utopia_m12h24 m10_h20 1.1250
#> 347  utopia_m12h24 m11_h20 1.2500
#> 348  utopia_m12h24 m12_h20 1.5000
#> 349  utopia_m12h24 m01_h21 1.3200
#> 350  utopia_m12h24 m02_h21 1.2650
#> 351  utopia_m12h24 m03_h21 1.1000
#> 352  utopia_m12h24 m04_h21 0.9900
#> 353  utopia_m12h24 m05_h21 0.9900
#> 354  utopia_m12h24 m06_h21 1.1000
#> 355  utopia_m12h24 m07_h21 1.2100
#> 356  utopia_m12h24 m08_h21 1.2100
#> 357  utopia_m12h24 m09_h21 1.0450
#> 358  utopia_m12h24 m10_h21 0.9900
#> 359  utopia_m12h24 m11_h21 1.1000
#> 360  utopia_m12h24 m12_h21 1.3200
#> 361  utopia_m12h24 m01_h22 1.1400
#> 362  utopia_m12h24 m02_h22 1.0925
#> 363  utopia_m12h24 m03_h22 0.9500
#> 364  utopia_m12h24 m04_h22 0.8550
#> 365  utopia_m12h24 m05_h22 0.8550
#> 366  utopia_m12h24 m06_h22 0.9500
#> 367  utopia_m12h24 m07_h22 1.0450
#> 368  utopia_m12h24 m08_h22 1.0450
#> 369  utopia_m12h24 m09_h22 0.9025
#> 370  utopia_m12h24 m10_h22 0.8550
#> 371  utopia_m12h24 m11_h22 0.9500
#> 372  utopia_m12h24 m12_h22 1.1400
#> 373  utopia_m12h24 m01_h23 0.9600
#> 374  utopia_m12h24 m02_h23 0.9200
#> 375  utopia_m12h24 m03_h23 0.8000
#> 376  utopia_m12h24 m04_h23 0.7200
#> 377  utopia_m12h24 m05_h23 0.7200
#> 378  utopia_m12h24 m06_h23 0.8000
#> 379  utopia_m12h24 m07_h23 0.8800
#> 380  utopia_m12h24 m08_h23 0.8800
#> 381  utopia_m12h24 m09_h23 0.7600
#> 382  utopia_m12h24 m10_h23 0.7200
#> 383  utopia_m12h24 m11_h23 0.8000
#> 384  utopia_m12h24 m12_h23 0.9600
#> 385 utopia_seasons WIN_DAY 1.2000
#> 386 utopia_seasons SPR_DAY 0.9000
#> 387 utopia_seasons SUM_DAY 1.1000
#> 388 utopia_seasons AUT_DAY 0.9000
#> 389 utopia_seasons WIN_NGT 0.7200
#> 390 utopia_seasons SPR_NGT 0.5400
#> 391 utopia_seasons SUM_NGT 0.6600
#> 392 utopia_seasons AUT_NGT 0.5400
#> 393 utopia_seasons  WIN_PK 1.6200
#> 394 utopia_seasons  SPR_PK 1.2150
#> 395 utopia_seasons  SUM_PK 1.4850
#> 396 utopia_seasons  AUT_PK 1.2150